[tests] Separate tests for simple lines; minor fixes
