+++ run-docker-tests.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 pushd /usr/src/
 xvfb-run nose2-3
+#xvfb-run nose2-3 tests.test_pg_source_simple_polygons
+#xvfb-run nose2-3 tests.test_table_table
 #xvfb-run nose2-3 tests.test_pg_table_pg_table
 #xvfb-run nose2-3 tests.test_table_table
 #xvfb-run nose2-3 tests.test_source_simple_polygons
