+++ tests/test_source_simple_polygons.py
@@ -63,42 +63,5 @@ class TestSimplePolySimplePoly(unittest.TestCase):
         self.plugin.unload()
 
 
-class TestSimplePolySimpleLine(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(self):
-        print('\nINFO: Set up simple_pol-simple_lin')
-        from AppendFeaturesToLayer.append_features_to_layer_plugin import AppendFeaturesToLayerPlugin
-        self.plugin = AppendFeaturesToLayerPlugin(get_iface)
-        self.plugin.initGui()
-        self.common = CommonTests()
-
-    def test_copy_all(self):
-        print('\nINFO: Validating simple_pol-simple_lin copy&paste all...')
-        output_layer, layer_path = get_qgis_gpkg_layer('target_simple_lines')
-        res = self.common._test_copy_all('source_simple_polygons', output_layer, layer_path)
-        layer = res['TARGET_LAYER']
-        self.assertEqual(layer.featureCount(), 1)
-
-    def test_copy_selected(self):
-        print('\nINFO: Validating simple_pol-simple_lin copy&paste selected...')
-        output_layer, layer_path = get_qgis_gpkg_layer('target_simple_lines')
-        res = self.common._test_copy_selected('source_simple_polygons', output_layer, layer_path)
-        layer = res['TARGET_LAYER']
-        self.assertEqual(layer.featureCount(), 0)  # Selected id has a hole, cannot be copied
-        self.assertEqual(res[APPENDED_COUNT], 0)
-
-        output_layer, layer_path = get_qgis_gpkg_layer('target_simple_lines')
-        res = self.common._test_copy_selected('source_simple_polygons', output_layer, layer_path, select_id=2)
-        layer = res['TARGET_LAYER']
-        self.assertEqual(layer.featureCount(), 1)
-        self.assertEqual(res[APPENDED_COUNT], 1)
-
-    @classmethod
-    def tearDownClass(self):
-        print('INFO: Tear down simple_pol-simple_lin')
-        self.plugin.unload()
-
-
 if __name__ == '__main__':
     nose2.main()
